Add merge operations
